make sure to clean state on close
